fix(fcbc): build the unit-test node helper with the u8 arity field
`ExpressionNode::arity` is a u8; the hand-pushed node helper in the
new unit test cast the operand count to u16 and failed the lib test
build (E0308). Count the present operands as u8.

+++ crates/fcs-fcbc/src/evaluator.rs
@@ -1353,7 +1353,7 @@ mod tests {
                     arity: operands
                         .iter()
                         .filter(|operand| **operand != u32::MAX)
-                        .count() as u16,
+                        .count() as u8,
                     immediate: 0,
                 });
                 node
